requirements.txt
Flask>=2.0
requests
piper-tts==1.2.0
Flask-Cors
PyInstaller